comments have been addressed
