+++ client/src/pages/Messages/ChatBox.tsx
@@ -52,14 +52,8 @@ const ChatBox = ({ profileId }: Props): JSX.Element => {
   };
 
   const handleSendMessage = () => {
-    const AsyncSendMessage = async () => {
-      if (
-        currentConversation &&
-        currentRecipient &&
-        currentConversation.conversationId &&
-        currentRecipient._id &&
-        socket
-      ) {
+    const asyncSendMessage = async () => {
+      if (currentConversation?.conversationId && currentRecipient?._id && socket) {
         const conversationId = currentConversation.conversationId;
         const recipientId = currentRecipient._id;
 
@@ -78,7 +72,7 @@ const ChatBox = ({ profileId }: Props): JSX.Element => {
             replyAreaRef.current.value = '';
           }
 
-          const setCurrentConversation = async () => {
+          {
             const response = await getOneConversation(conversationId);
 
             if (response.error) {
@@ -86,14 +80,12 @@ const ChatBox = ({ profileId }: Props): JSX.Element => {
             } else if (response.conversationContent) {
               updateCurrentConversation(response.conversationContent);
             }
-          };
-
-          setCurrentConversation();
+          }
         }
       }
     };
 
-    AsyncSendMessage();
+    asyncSendMessage();
   };
 
   const handleSendTextChange = (e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
