+++ server/app.js
@@ -17,9 +17,10 @@ const dogRouter = require("./routes/dog");
 const paymentRouter = require("./routes/stripe");
 const conversationRouter = require("./routes/conversation");
 const messageRouter = require("./routes/message");
-const { addUser, removeUser, getUser } = require("./utils/socketUsers");
+
 const ioCookieParser = require("socket.io-cookie-parser");
 const jwt = require("jsonwebtoken");
+const { authentication, messaging } = require("./middleware/messagingSocket");
 
 const { json, urlencoded } = express;
 
@@ -33,37 +34,9 @@ const io = socketio(server, {
   },
 });
 io.use(ioCookieParser());
-io.use((socket, next) => {
-  const token = socket.request.cookies["token"];
-  if (token) {
-    try {
-      const decoded = jwt.verify(token, process.env.JWT_SECRET);
-      user = decoded;
-      next();
-    } catch (err) {
-      return next(new Error("Authentication Error"));
-    }
-  } else {
-    return next(new Error("Authentication Error"));
-  }
-}).on("connection", (socket, next) => {
-  socket.on("addUser", (userId) => {
-    const users = addUser(userId, socket.id);
-    io.emit("getUsers", users);
-  });
-
-  socket.on("sendMessage", ({ recipientId, ...rest }) => {
-    const activeUser = getUser(recipientId);
-    if (activeUser) {
-      io.to(activeUser.socketId).emit("getMessage", { ...rest });
-    }
-  });
-
-  socket.on("disconnect", () => {
-    const users = removeUser(socket.id);
-    io.emit("getUsers", users);
-  });
-});
+io.use(authentication).on("connection", (socket, next) =>
+  messaging(socket, next, io)
+);
 
 if (process.env.NODE_ENV === "development") {
   app.use(logger("dev"));
