+++ server/controllers/user.js
@@ -1,44 +1,6 @@
 const User = require("../models/User");
 const asyncHandler = require("express-async-handler");
-/**
- * 
- *   const profiles = await Profile.find()
-    .populate({
-      path: "user",
-      select: { password: 0 },
-      match: {
-        role: "sitter",
-      },
-      options: { retainNullValues: false },
-      transform: (doc, id) => (doc == null ? null : doc),
-    })
-    .exec()
-    .then((profiles) => {
-      return profiles.filter((profile) => {
-        if (filter) {
-          if (profile.user) {
-            if (
-              !_.isEmpty(profile.availability[dateIn]) &&
-              !_.isEmpty(profile.availability[dateOff])
-            ) {
-              const searchTarget = [
-                profile.firstName.toLowerCase(),
-                profile.lastName.toLowerCase(),
-                profile.address.toLowerCase(),
-              ];
-              const result = searchTarget.find((searchTargetItem) =>
-                searchTargetItem.includes(search.toLowerCase())
-              );
-              return !!result;
-            }
-          }
-          return false;
-        } else {
-          return profile.user != null;
-        }
-      });
-    });
- */
+
 // @route POST /users
 // @desc Search for users
 // @access Private
@@ -52,10 +14,8 @@ exports.searchUsers = asyncHandler(async (req, res, next) => {
     })
       .select(" -password  ")
       .populate("profile", "firstName lastName")
-      .exec()
-      .then((result) => {
-        return result.filter((userWithProfile) => !!userWithProfile.profile);
-      });
+      .exec();
+    users = await users.filter((userWithProfile) => !!userWithProfile.profile);
   }
   if (!users) {
     res.status(404);
